+++ decidim-assemblies/lib/decidim/assemblies/admin_engine.rb
@@ -46,6 +46,12 @@ module Decidim
               put :hide
             end
           end
+
+          resources :participatory_space_private_users, controller: "participatory_space_private_users" do
+            member do
+              post :resend_invitation, to: "participatory_space_private_users#resend_invitation"
+            end
+          end
         end
 
         scope "/assemblies/:assembly_slug/components/:component_id/manage" do