+++ decidim-participatory_processes/app/models/decidim/participatory_process.rb
@@ -15,6 +15,7 @@ module Decidim
     include Decidim::Followable
     include Decidim::HasReference
     include Decidim::Traceable
+    include Decidim::HasPrivateUsers
     include Decidim::Loggable
     include Decidim::ParticipatorySpaceResourceable
 
@@ -75,6 +76,10 @@ module Decidim
 
     def to_param
       slug
+    end
+
+    def self.private_processes
+      where(private_space: true)
     end
   end
 end