+++ decidim-proposals/spec/system/vote_proposal_spec.rb
@@ -92,7 +92,7 @@ describe "Vote Proposal", type: :system do
           end
 
           within "#proposal-#{proposal.id}-votes-count" do
-            expect(page).to have_content("1/0 VOTE")
+            expect(page).to have_content("1 VOTE")
           end
         end
       end
@@ -110,7 +110,7 @@ describe "Vote Proposal", type: :system do
           end
 
           within "#proposal-#{proposal.id}-votes-count" do
-            expect(page).to have_content("1/0 VOTE")
+            expect(page).to have_content("1 VOTE")
           end
         end
 
@@ -260,7 +260,7 @@ describe "Vote Proposal", type: :system do
             end
 
             it "shows the vote count but not the vote button" do
-              expect(page).to have_css(".card__support__data", text: "1/0 VOTE")
+              expect(page).to have_css(".card__support__data", text: "1 VOTE")
               expect(page).to have_content("Voting disabled")
             end
           end
@@ -288,7 +288,7 @@ describe "Vote Proposal", type: :system do
       let!(:component) do
         create(:proposal_component,
                :with_votes_enabled,
-               :with_maximum_votes_per_proposal,
+               :with_threshold_per_proposal,
                manifest: manifest,
                participatory_space: participatory_process)
       end
@@ -327,7 +327,7 @@ describe "Vote Proposal", type: :system do
       let!(:component) do
         create(:proposal_component,
                :with_votes_enabled,
-               :with_maximum_votes_per_proposal,
+               :with_threshold_per_proposal,
                :with_can_accumulate_supports_beyond_threshold,
                manifest: manifest,
                participatory_space: participatory_process)
@@ -345,7 +345,7 @@ describe "Vote Proposal", type: :system do
 
         within proposal_element do
           within ".card__support", match: :first do
-            expect(page).to have_content("1/0 VOTE")
+            expect(page).to have_content("1 VOTE")
           end
         end
       end